Implemented flush boolean to the ChatPacket SendServerMessage function overload Added message to inform the user when a command isn't implemented Added more comments
